Migrated to HassOS 4.9 and my zwave/zigbee usb dongles work again
This was a pain. Still not sure what happened. But the upgrade to HassOS 4.8 killed both
zwave2mqtt and zigbee2mqtt is seems. After the upgrade neitehr addon would talk to the hardware.
I migrated back to smartthings...major pain..and then after 4.10 was released was able to get the
addons added abck and talking to the hardware again.

All lights and switches are back talking locally. Also, it takes about 18 hours to migrate back to smartthings,
and about 18 hours to migrate back to local control.

Let's hope that doesn't happen again.